requirements.txt
################################################################################
# This requirements file has been automatically generated from `Pipfile` with
# `pipenv-to-requirements`
#
#
# This has been done to maintain backward compatibility with tools and services
# that do not support `Pipfile` yet.
#
# Do NOT edit it directly, use `pipenv install [-d]` to modify `Pipfile` and
# `Pipfile.lock` and then regenerate `requirements*.txt`.
################################################################################

alembic==1.12.1 ; python_version >= '3.7'
appdirs==1.4.4
azure-core==1.29.5 ; python_version >= '3.7'
azure-storage-blob==12.19.0
bcrypt==4.0.1 ; python_version >= '3.6'
beautifulsoup4==4.12.2 ; python_full_version >= '3.6.0'
blinker==1.7.0 ; python_version >= '3.8'
bs4==0.0.1
certifi==2023.7.22 ; python_version >= '3.6'
cffi==1.16.0 ; python_version >= '3.8'
charset-normalizer==3.3.2 ; python_full_version >= '3.7.0'
click==8.1.7 ; python_version >= '3.7'
cryptography==41.0.5 ; python_version >= '3.7'
cssselect==1.2.0 ; python_version >= '3.7'
dateparser==1.1.8 ; python_version >= '3.7'
fake-useragent==1.3.0
feedfinder2==0.0.4
feedparser==6.0.10 ; python_version >= '3.6'
filelock==3.13.1 ; python_version >= '3.8'
flask-bcrypt==1.0.1
flask-cors==4.0.0
flask-jwt-extended==3.18.0
flask-jwt==0.3.2
flask-login==0.6.3
flask-migrate==4.0.5
flask-sqlalchemy==3.1.1
flask==3.0.0
googlenews==1.6.10
greenlet==3.0.1 ; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
idna==3.4 ; python_version >= '3.5'
importlib-metadata==6.8.0 ; python_version < '3.10'
importlib-resources==6.1.1 ; python_version < '3.10'
isodate==0.6.1
itsdangerous==2.1.2 ; python_version >= '3.7'
jieba3k==0.35.1
jinja2==3.1.2 ; python_version >= '3.7'
joblib==1.3.2 ; python_version >= '3.7'
lxml==4.9.3 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
mako==1.3.0 ; python_version >= '3.8'
markupsafe==2.1.3 ; python_version >= '3.7'
newspaper3k==0.2.8
nltk==3.8.1 ; python_version >= '3.7'
parse==1.19.1
pillow==10.1.0 ; python_version >= '3.8'
psycopg2-binary==2.9.9
pycparser==2.21
pyee==8.2.2
pyjwt==1.4.2
pyppeteer==1.0.2 ; python_version >= '3.7' and python_version < '4.0'
pyquery==2.0.0
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.0
pytz==2023.3.post1
pyyaml==6.0.1 ; python_version >= '3.6'
regex==2023.10.3 ; python_version >= '3.7'
requests-file==1.5.1
requests-html==0.10.0
requests==2.31.0
sgmllib3k==1.0.0
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
soupsieve==2.5 ; python_version >= '3.8'
sqlalchemy==2.0.23
tinysegmenter==0.3
tldextract==5.1.0 ; python_version >= '3.8'
tqdm==4.66.1 ; python_version >= '3.7'
typing-extensions==4.8.0 ; python_version >= '3.8'
tzlocal==5.2 ; python_version >= '3.8'
urllib3==1.26.18 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
w3lib==2.1.2 ; python_version >= '3.7'
websockets==10.4 ; python_version >= '3.7'
werkzeug==3.0.1 ; python_version >= '3.8'
zipp==3.17.0 ; python_version >= '3.8'
